Added playback from file

+++ core/src/commonMain/kotlin/AudioPlayer.kt
@@ -1,7 +1,7 @@
 package pl.lemanski.pandaloop
 
 expect object AudioPlayer {
-    fun initializePlaybackDevice(buffer: FloatArray)
+    fun initializePlaybackDevice(buffer: ByteArray)
     fun uninitalizePlaybackDevice()
     fun startPlayback()
     fun stopPlayback()
